Improve PromQL parser performance by making it non-concurrent
Before this commit, the PromQL parser ran in two goroutines:
* The lexer goroutine that splits the input into tokens and sent them over a channel to
* the parser goroutine which produces the abstract syntax tree

The Problem with this approach is that the parser spends more time on goroutine creation
and syncronisation than on actual parsing.

This commit removes that concurrency and replaces the channel by a slice based buffer.

Benchmarks show that this makes the parser 3-5 times as fast as before.

Signed-off-by: Tobias Guggenmos <tguggenm@redhat.com>